requirements.txt
aiosmtplib==4.0.0
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
h11==0.14.0
idna==3.10
isort==6.0.1
motor==3.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.3
python-dotenv==1.0.1
ruff==0.11.0
sniffio==1.3.1
starlette==0.46.1
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt